4_observability/sk_backend/pyproject.toml
[project]
name = "sk-orchestrator-backend"
version = "0.1.0"
description = "Semantic Kernel + FastAPI backend for beauty product chatbot with XML-grounded context."
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "semantic-kernel[azure]==1.41.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "azure-identity>=1.18.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "azure-monitor-opentelemetry>=1.6.0",
]

[project.scripts]
sk-orchestrator = "sk_orchestrator.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
